change debug command to show-pr
Use a more descriptive name and improve the argument handling to make
the UI nicer.
